Split out reply-parsing code
